+++ tests/test_versioned_message.py
@@ -334,3 +334,4 @@ def test_try_compile() -> None:
             )
         ],
     )
+